drivers/mmc/core/core.c: Disable SDIO and SD card detection steps, updated sdhxi-xlp driver
Micron eMMC card is misbehaving when receiving non-MMC commands as part
of the SDIO and SD card identification sequences that are performed by
the core code prior to identifying the MMC device. For now, just disabling
those steps. Next step is to figure out why the eMMC device is misbehaving
and address it.